#3694 Fix dashboard table maximum results

+++ stroom-core-shared/src/main/java/stroom/dashboard/shared/TableComponentSettings.java
@@ -48,7 +48,7 @@ import java.util.Objects;
 @JsonInclude(Include.NON_NULL)
 public class TableComponentSettings implements ComponentSettings {
 
-    public static final int[] DEFAULT_MAX_RESULTS = {1000000};
+    public static final long[] DEFAULT_MAX_RESULTS = {1000000L};
 
     @Schema(description = "TODO", required = true)
     @JsonProperty
@@ -75,7 +75,7 @@ public class TableComponentSettings implements ComponentSettings {
             "'1000,10,1' means 1000 results at group level 0, 10 at level 1 and 1 at level 2. " +
             "In the absence of this field system defaults will apply")
     @JsonProperty
-    private final List<Integer> maxResults;
+    private final List<Long> maxResults;
 
     @Schema(description = "Defines the maximum number of rows to display in the table at once (default 100).",
             example = "100")
@@ -103,7 +103,7 @@ public class TableComponentSettings implements ComponentSettings {
             @JsonProperty("fields") final List<Field> fields,
             @JsonProperty("extractValues") final Boolean extractValues,
             @JsonProperty("extractionPipeline") final DocRef extractionPipeline,
-            @JsonProperty("maxResults") final List<Integer> maxResults,
+            @JsonProperty("maxResults") final List<Long> maxResults,
             @JsonProperty("pageSize") final Integer pageSize,
             @JsonProperty("showDetail") final Boolean showDetail,
             @JsonProperty("conditionalFormattingRules") final List<ConditionalFormattingRule>
@@ -149,7 +149,7 @@ public class TableComponentSettings implements ComponentSettings {
         return extractionPipeline;
     }
 
-    public List<Integer> getMaxResults() {
+    public List<Long> getMaxResults() {
         return maxResults;
     }
 
@@ -247,7 +247,7 @@ public class TableComponentSettings implements ComponentSettings {
         private List<Field> fields;
         private Boolean extractValues;
         private DocRef extractionPipeline;
-        private List<Integer> maxResults;
+        private List<Long> maxResults;
         private Integer pageSize;
         private Boolean showDetail;
         private List<ConditionalFormattingRule> conditionalFormattingRules;
@@ -353,7 +353,7 @@ public class TableComponentSettings implements ComponentSettings {
             return this.extractionPipeline(DocRef.builder().type(type).uuid(uuid).name(name).build());
         }
 
-        public Builder maxResults(final List<Integer> maxResults) {
+        public Builder maxResults(final List<Long> maxResults) {
             this.maxResults = maxResults;
             return this;
         }
