+++ stroom-legacy/stroom-legacy-impex_6_1/src/main/java/stroom/legacy/impex_6_1/MappingUtil.java
@@ -184,10 +184,10 @@ public final class MappingUtil {
             return null;
         }
 
-        List<Integer> maxResults = null;
+        List<Long> maxResults = null;
         if (value.getMaxResults() != null) {
             maxResults = new ArrayList<>();
-            for (int results : value.getMaxResults()) {
+            for (long results : value.getMaxResults()) {
                 maxResults.add(results);
             }
         }
